+++ src/components/common/AppLayout/AppLayout.tsx
@@ -1,5 +1,5 @@
-import Footer from "../Footer";
 import Header from "../Header";
+import Footer from "../Footer";
 import React from "react";
 import * as S from "./style";
